moved velocity_models module into subpackage
