+++ halotools/empirical_models/__init__.py
@@ -19,4 +19,3 @@ from .hod_designer import *
 from .phase_space_models import *
 # from .profile_helpers import *
 from .profile_models import *
-from .velocity_models import *
